isolate logging setup to middlewares
