+++ src/middlewares/logging.js
@@ -1,0 +1,18 @@
+/**
+ * app logging
+ */
+
+const winston = require('winston');
+
+
+const logger = winston.createLogger({
+    level: 'debug',
+    format: winston.format.json(),
+    transports: [
+        new winston.transports.File({ filename: 'logs/error.log', level: 'error' }),
+        new winston.transports.File({ filename: 'logs/combined.log' }),
+    ],
+});
+
+
+module.exports = logger;
